+++ examples/udp_sender_receiver/udp_sender.c
@@ -149,7 +149,7 @@ int main(int argc, char* argv[])
         exit(1);
     }
 
-    uint32_t trace_enabled = 0;
+    uint8_t trace_enabled = 0;
     // Create the encoder factory
     encoder_factory = kodo_new_encoder_factory(algorithm, finite_field,
                                                max_symbols, max_symbol_size,
